Removing stub doc files
There are a ton of wishful-thinking traslation subdirs under docs/
which contain nothing but files which say "Please see..." and
directs to the English version.

Removing them, because they're just cluttering thigns,
and it makes it hard to tell whether any translations actually _do_
exist (which may require attention; e.g. revising for a new
release).