FIXME note in Makefile: starter/template thumbs...
The thumbnails for starters & templates are NOT being re-generated
when the source images are modified -- only when the thumbnail is
missing.  Needs an update to the target's prerequisites
(but I'm very rusty with this level of Makefile magic).

Also, update some Starter source images so they work better with
flood fill (Bald Eagle, World map, Gecko).