Wrap up GIF export UI
GIF export button now jumps to a process that iterates through the
chosen saved images, loading each (but not yet saving as a GIF).
It then returns to the Open->Slideshow dialog, and displays a
pop-up dialog that shows a success/fail message.

Also, adding an icon for a regular static image "Export" option,
to be added to the main Open dialog.