WIP: Shape controls -- center vs corner
Adding control buttons to the Shapes tool, allowing for shapes
to be drawn from the center (as before) or from a corner
(more like other paint packages).  The controls do nothing
at this time, but are visible and can be clicked.  This is a
work in progress.

The controls may be removed using a configuration option.