Keyboard max size didn't account for UI buttons
Keyboard's goal is to not be > 90% of the width of the drawing
canvas.  Was misinterpreting the incoming argument ('canvas'),
so have to account for the toolbar (left) and selector (right)
buttons in the UI.

So e.g., `tuxpaint --900x768 --onscreen-keyboard` will use the
small keyboard buttons, while `... --1024x768 ...` will use the large.