Ignore & clean up failed SVG thumbnails
When building Tux Paint, ImageMagick's `convert` might fail to generate
thumbnail PNGs for some Starter SVG images (e.g., "manatee.svg"),
with a "too many bezier coordinates" error.

This did not happen before, and does not happen with newer versions
of ImageMagick; there was a regression fixed in 2018:
https://github.com/ImageMagick/ImageMagick/issues/1069

However, it's occurring in Ubuntu 20.04 'focal' (I've got
version "6.9.10-23 Q16 x86_64 20190101").  Therefore, as a quick-fix,
simply ignoring errors from `convert`, and removing the (bogus)
PNG thumbmail image, and allowing the `make` process to proceed.

Tux Paint will simply open the SVG and generate a thumbnail to
display on the screen when the "New" dialog is brought up.