Pulling in initial English from tuxpaint-docs
Documentation is now maintained in a separate source repository,
'tuxpaint-docs', which allows for easier and more consisten
translation of the documentation.  Pulling in initial version of
these docs, which contain a variety of minor updates (and also a
few which had previously only been available as plain TXT are now
available as HTML).

Pulling in the English stuff.